pyproject.toml
[tool.poetry]
name = "cowboy"
version = "0.1.0"
description = ""
authors = ["John Peng"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "1.14.0"
anyio = "4.6.2.post1"
black = "24.10.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
click = "8.1.7"
distro = "1.9.0"
dnspython = "2.7.0"
docstring-parser = "0.16"
ecdsa = "0.19.0"
email-validator = "2.2.0"
fastapi = "0.115.5"
gitdb = "4.0.11"
gitpython = "3.1.43"
greenlet = "3.1.1"
h11 = "0.14.0"
httpcore = "1.0.7"
httpx = "0.27.2"
idna = "3.10"
jiter = "0.7.1"
mako = "1.3.6"
markupsafe = "3.0.2"
mypy-extensions = "1.0.0"
openai = "1.55.1"
packaging = "24.2"
pathspec = "0.12.1"
platformdirs = "4.3.6"
psycopg2 = "2.9.10"
psycopg2-binary = "2.9.10"
pyasn1 = "0.6.1"
pydantic = "2.9.1"
python-jose = "3.3.0"
regex = "2024.11.6"
requests = "2.32.3"
rsa = "4.9"
sentry-sdk = {extras = ["fastapi"], version = "2.19.0"}
simple-parsing = "0.1.6"
six = "1.16.0"
smmap = "5.0.1"
sniffio = "1.3.1"
sqlalchemy = "2.0.36"
starlette = "0.41.3"
tenacity = "9.0.0"
tiktoken = "0.8.0"
tqdm = "4.67.1"
typing-extensions = "4.12.2"
urllib3 = "2.2.3"
uvicorn = "0.32.1"
pytz = "^2024.2"
bcrypt = "^4.2.1"
boto3 = "^1.35.70"
jinja2 = "^3.1.4"
braintrust = "^0.0.175"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"